Rename constraint applyer to constraint applier
